backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.7
pypdf==3.17.1
chromadb==0.4.18
tiktoken==0.5.2
structlog==23.2.0
